get_extensions: use shutil.copyfile to avoid PermissionError
Without this change, packages that depend on this module (e.g astropy)
may get a PermissionError on systems such as Nix and Guix. Because, when
these systems attempt to build astropy, the built package
`astropy-extrnsion-helpers` holds an `src/compiler.c` that has no write
permissions - just like any other file in a built package in these
systems. Then, when astropy tries to use the copied `_compiler.c`, it
tries to open it with `wb+`. Using `shutil.copyfile` doesn't copy
metadata.